+++ src/Common/model/search/SearchIndex.swift
