+++ src/Common/model/timeseries/TimeSeriesModel.swift
