+++ src/Common/model/timeseries/TimeSeriesValue.swift
